refactor(2d,sketching): extract base sketchers to Layer 2, fix boundary
Layer 3 audit F1+F2 — `sketcher2d.ts` (692 LOC) bundled three classes
with conflicting layer requirements:
- BaseSketcher2d: pure 2D pen-drawing engine (no Layer 3 deps)
- BlueprintSketcher: returns Blueprint (a Layer 2 type)
- FaceSketcher: returns Sketch (a Layer 3 type)

The first two had no business being in Layer 3 at all, which is why
`cannedBlueprints.ts` (Layer 2) had to import from `@/sketching/sketcher2d.js`
(Layer 3) — an acknowledged boundary violation whitelisted in
`scripts/check-layer-boundaries.sh`.

Move the layer-2 pieces to where they belong:
- `src/2d/blueprints/baseSketcher2d.ts` — BaseSketcher2d
- `src/2d/blueprints/blueprintSketcher.ts` — BlueprintSketcher
- `src/2d/blueprints/genericSketcher.ts` — was sketcherlib.ts
- `src/2d/blueprints/ellipseUtils.ts` — was sketching/ellipseUtils.ts

Split the layer-3 piece into its own file:
- `src/sketching/faceSketcher.ts` — FaceSketcher (still Layer 3, returns Sketch)

Delete `src/sketching/sketcher2d.ts` and remove the boundary-checker
whitelist. All public exports preserved via updated re-exports in
`src/index.ts` and `src/sketching.ts`.
